MAJOR COMMIT: correction/retract/replace
This commit adds months of changes to this repository. In March 2023,
it was brought to my attention that one of the key results in the paper
was a bit odd. This lead me to double-check everything, and ultimately
discovered a typo in the wBarY() function in the R simulation code
file ./R/simulations-inversions-delMut.R. I initiated the process
of a correction with Evolution, which became a retract/replace.

Over the following months, I made a variety of changes to the code
and repo, including completely re-coding almost all of the key fuct
ions and recursions. It also involved deriving two new models, and
updating the paper significantly. In this commit, I am finally updat
ing the GitHub repo. The major changes include:

- New files with updated/corrected code, all marked with *-RECODE.R
- New files: R/Auto-Inversions-Nei-1967.R and
			 R/Auto-Inversions-NeiMyVersion.R
			 which include code developed to reproduce and extend the
			 model of Net et al. (1967) for autosomal inversions.
- Updated code for making figures